+++ react/ProductQuantity.tsx
@@ -20,9 +20,11 @@ const ProductQuantity: StorefrontFunctionComponent<Props>  = ({
 
 interface Props {
   warningQuantityThreshold: number
-  selectedQuantity: number
-  selectedItem: any
-  dispatch: any
+}
+
+ProductQuantity.schema = {
+  title: 'admin/editor.product-quantity.title',
+  description: 'admin/editor.product-quantity.description',
 }
 
 export default ProductQuantity